// client/src/components/molecules/ServiceCard.d.ts
import React from 'react';
interface ServiceCardProps {
    title: string;
    description: string;
    icon: string;
    image: string;
    link: string;
}
/**
 * ServiceCard Component
 *
 * Displays a service with an icon, image, title, description, and a link to the service details page.
 * Used on the homepage and services page to showcase the different services offered.
 */
declare const ServiceCard: React.FC<ServiceCardProps>;
export default ServiceCard;
//# sourceMappingURL=ServiceCard.d.ts.map